Power series Newton's method
We prove the general case of one-dimensional power series Newton's
method correct, and test it with one initial case: series inverses.
Adding square roots is the next step, and then we have the basic
machinery we need for Bottcher series computations.

This proved surprisingly finicky, and for somewhat silly reasons: there
is poor automation (at least to my abilities) for `WithTop ℕ∞`
inequalities, and the precise degree of continuous differentiability is
awkward to get right. My sense is that grind could handle the former if
I learned how to use it well. The latter could have been simplified by
requiring infinite smoothness throughout, which is all I need for
applications, but it's sort of neat to have the arbitrary finite case.

Finally, I burned a bunch of time on proving that two different power
series solutions to a nonlinear equation are correct, before finally
settling on a nice approach as implemented in Unique.lean.